XMLHttpRequest/open-url-base-inserted-after-open.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() resolving URLs - insert &lt;base> after open()</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest(),
            base = document.createElement("base")
        base.href = "http://localhost/xhr/resources/"
        client.open("GET", "folder.txt", false)
        document.getElementsByTagName("head")[0].appendChild(base)
        client.send(null)
        assert_equals(client.responseText, "top\n")
      })
    </script>
  </body>
</html>
